[#1] rozbicie_na_pliki
