+++ db_sqlite_fastapi/place_sqlite.py
@@ -1,0 +1,21 @@
+#SqlAlchemy Setup
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker, declarative_base
+
+SQLALCHEMY_DATABASE_URL = 'sqlite+pysqlite:///.db.sqlite3:'
+engine = create_engine(SQLALCHEMY_DATABASE_URL, echo=True, future=True)
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+Base = declarative_base()
+
+
+# Depedency
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+
+Base.metadata.create_all(bind=engine)
+
