+++ fastapi_validate_data_pydantic.py
@@ -1,0 +1,36 @@
+from typing import Optional
+
+from pydantic import BaseModel
+from sqlalchemy.orm import Session
+
+from fastapi_db_sqlite import DBPlace
+
+
+class Place(BaseModel):
+    name: str
+    description: Optional[str] = None
+    coffee: bool
+    wifi: bool
+    food: bool
+    lat: float
+    lng: float
+
+    class Config:
+        orm_mode = True
+
+
+def get_place(db: Session, place_id: int):
+    return db.query(DBPlace).where(DBPlace.id == place_id).first()
+
+
+def get_places(db: Session):
+    return db.query(DBPlace).all()
+
+
+def create_place(db: Session, place: Place):
+    db_place = DBPlace(**place.dict())
+    db.add(db_place)
+    db.commit()
+    db.refresh(db_place)
+
+    return db_place
